Refactor: Move methods from EmployeeService to TransactionService & Implement Employee Report
- Moved balance calculations and relevant methods from EmployeeService to TransactionService.
- Implemented `EmployeeReport` method in TransactionService to generate leave balance reports.
- Added `EmployeeReport` endpoint in the controller for fetching employee leave reports.
- Developed Balance Details Report endpoint for better transaction reporting.
