--- RequestsManagementSystem.DTOs/api/TransactionsDtos/TransactionForReportDTO.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RequestsManagementSystem.DTOs.api.TransactionsDtos
-{
-    public class TransactionForReportDTO
-    {
-        public string StartDate { get; set; }
-        public string EndDate { get; set; }
-        public string Duration { get; set; }
-    }
-}
