+++ RequestsManagementSystem.Logic/Services/IEmployeeService.cs
@@ -9,7 +9,6 @@ namespace RequestsManagementSystem.Logic.Services
         Task<EmployeeDto> GetEmployeeDataAsync(int id);
         Task<UpdatePasswordResultDto> UpdatePasswordAsync(UpdatePasswordEmployeeDto EmployeeDto);
         Task<IEnumerable<EmployeeIdAndNameDto>> GetEmployeesAsync(string departmentName);
-        (double CasualBalance, double RegularBalance) GetEmployeeBalance(Employee employee, DateOnly? p_startDate = null, DateOnly? p_endDate = null);
         Task<IEnumerable<EmployeeExcelDto>> GetEmployeesToExcelFormat(DateOnly? startDate, DateOnly? EndDate);
     }
 }
