+++ RequestsManagementSystem.Logic/Services/ITransactionService.cs
@@ -1,4 +1,4 @@
-﻿using RequestsManagementSystem.Core.Enums;
+﻿using RequestsManagementSystem.Core.Entities;
 using RequestsManagementSystem.DTOs.api;
 using RequestsManagementSystem.DTOs.api.TransactionsDtos;
 
@@ -15,6 +15,9 @@ namespace RequestsManagementSystem.Logic.Services
 		Task SetSeenStatus(int id, string whoSeen);
         Task<TransactionDto?> GetTransactionByIdAsync(int id);
         Task<string> UpdateTransactionStatusAsync(int id, UpdateTransactionStatusDto request);
-
+        int CalculateMonthCount(DateOnly employmentDate, DateOnly startDate, DateOnly endDate);
+        double CalculateLeaveInMonthRange(double leavesPerMonth, DateOnly employementDate, DateOnly p_startdate, DateOnly p_endDate);
+        (double CasualBalance, double RegularBalance) GetEmployeeBalance(Employee employee, DateOnly? p_startDate = null, DateOnly? p_endDate = null);
+        Task<ReportTransactionDTO> EmployeeReport(int EmployeeId, string p_type, DateTime? StartDate, DateTime? EndDate);
     }
 }
